% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_fc}
\alias{calculate_fc}
\title{Calculate a bikes front center length}
\usage{
calculate_fc(geo)
}
\arguments{
\item{tibble}{geo dataframe containing wheelbase and rear center
measurements}
}
\description{
Using wheelbase and rear center measurements, calculate the front center
length of a bike. The front center is defined as the distance from the center
of the bottom bracket, to the center of the front axle.
}
