% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate_fc_rc_ratio_static}
\alias{calculate_fc_rc_ratio_static}
\title{Calculate a bikes front to rear center static ratio}
\usage{
calculate_fc_rc_ratio_static(geo)
}
\arguments{
\item{tibble}{geo dataframe containing front and rear center measurements}
}
\description{
Using front and rear center measurements, calculate the front to rear center
ratio. This calculation assumes the bike is static, with zero sag in the
suspension, on flat ground.
}
