% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geo}
\alias{geo}
\title{Mountain bike geometry table}
\format{
A data frame of class "tbl" with 15 rows and 3 variables:
\describe{
\item{bike}{Bike model and size}
\item{wb}{Wheelbase}
\item{rc}{Rear center (chainstay) length, calculated from the center of the
bottom bracket, to the centre of the rear dropout}
...
}
}
\usage{
geo
}
\description{
A data set containing mountain bike geometry data, manually scraped from the
web.
}
\keyword{datasets}
